+++ nameui/ch3-ch5/src/main/java/controller/etc/CssController.java
@@ -10,7 +10,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-import static webserver.RequestHandler.BASE_URL;
+import static webserver.DispatcherServlet.BASE_URL;
 
 public class CssController implements Controller {
     @Override
