+++ nameui/ch3-ch5/src/main/java/controller/user/UserSaveController.java
@@ -12,7 +12,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-import static webserver.RequestHandler.BASE_URL;
+import static webserver.DispatcherServlet.BASE_URL;
 
 public class UserSaveController implements Controller {
     @Override
