+++ nameui/ch3-ch5/src/main/java/webserver/WebServer.java
@@ -26,8 +26,8 @@ public class WebServer {
             // 클라이언트가 연결될때까지 대기한다.
             Socket connection;
             while ((connection = listenSocket.accept()) != null) {
-                RequestHandler requestHandler = new RequestHandler(connection);
-                requestHandler.start();
+                DispatcherServlet dispatcherServlet = new DispatcherServlet(connection);
+                dispatcherServlet.start();
             }
         }
     }
